Version 3
Version 3 of the Rick's Guitar Shop case study is now implemented using Object-Oriented Programming principles in Java. In this updated version, we added new attributes to the GuitarSpec class to allow for searching guitars based on the number of strings. The search functionality in the Inventory class was also updated, with the implementation split between the Inventory and Guitar classes as a Boolean checker. Finally, the Guitar class was updated to implement encapsulation for the GuitarSpec class.